refactor: Added ava lint rules and fixed errors
Fixed issues:
- test/unit/util.js was being ignored - fixed
- test/util.js moved to test/_util.js to be implicitly ignored by ava
  instead of using an explicit ignore rule
- use `t.true` instead of `t.truthy` on boolean values
- move `test.after` hooks to be declared before tests are declared,
  ensuring they are actually run
