+++ package.json
@@ -24,6 +24,7 @@
   "dependencies": {
     "ajv": "^6.11.0",
     "email-validator": "^2.0.4",
+    "eslint-plugin-ava": "^10.2.0",
     "express": "^4.17.1",
     "mustache": "^4.0.0",
     "node-pg-migrate": "^4.2.2",
