+++ test/integration/submit.js
@@ -7,9 +7,16 @@ const db = require('../../server/database')
 const challenges = require('../../server/challenges')
 const { responseList } = require('../../server/responses')
 const auth = require('../../server/auth')
-const util = require('../util')
+const util = require('../_util')
 
 const chall = challenges.getAllChallenges()[0]
+
+test.after.always('remove solves from test user', async t => {
+  await db.solves.removeSolvesByUserId({ userid: uuid })
+  await db.auth.removeUserById({
+    id: testUser.id
+  })
+})
 
 test('fails with unauthorized', async t => {
   const resp = await request(app)
@@ -71,11 +78,4 @@ test.serial('fails with badAlreadySolvedChallenge', async t => {
     .expect(responseList.badAlreadySolvedChallenge.status)
 
   t.is(resp.body.kind, 'badAlreadySolvedChallenge')
-})
-
-test.after.always('remove solves from test user', async t => {
-  await db.solves.removeSolvesByUserId({ userid: uuid })
-  await db.auth.removeUserById({
-    id: testUser.id
-  })
 })
