+++ test/unit/challenges.js
@@ -5,5 +5,5 @@ const challenges = require('../../server/challenges')
 test('get all challenges', t => {
   const data = challenges.getAllChallenges()
 
-  t.truthy(Array.isArray(data))
+  t.true(Array.isArray(data))
 })
